+++ react-project/src/pages/Login.css
@@ -27,39 +27,20 @@ body {
     top: 55px;
     right: 40px;
     width: 10%;
-    
+
     display: flex;
     flex-direction: column;
     text-align: center;
     align-items: center;
     align-content: center;
     cursor: pointer;
-    
+
     padding: 0.5% 0.5%;
     border-radius: 5rem;
-    
+
     color: #000000;
     background-color: #f3d4c7;
     font-size: 1em;
-    letter-spacing: .15rem;
-    text-transform: uppercase;
-    font-weight: bold;
-}
-
-#signup-button, #signin-button {
-    text-align: center;
-    cursor: pointer;
-    align-items: center;
-    align-content: center;
-
-    margin-top: 20px;
-    padding: 5%;
-    width: 80%;
-    border-radius: 5rem ;
-
-    color: #000000;
-    background-color: #f3d4c7;
-    font-size: 1.5em;
     letter-spacing: .15rem;
     text-transform: uppercase;
     font-weight: bold;
@@ -81,44 +62,21 @@ h1 {
     text-align: center;
 }
 
-#signup-container, #login-container {
-    padding: 3.5em;
-    border-radius: 1rem;
-    box-shadow: 5px 15px 50px #303030;
-}
-
-#username-login-input, #password-login-input, 
-#email-signup-input, #username-signup-input,
-#password-signup-input {
-    border: 0;
-    border-bottom: 3px;
-    margin-top: 12px;
-    border-style: solid;
-    border-color: #dfdfdf;
-    border-radius: 1rem;
-    margin-bottom: 1.5em;
-
-    padding: 20px;
-    outline: none;
-    width: 20rem;
-
-    font-weight: 500;
-    font-size: 1em;
-}
-
 label {
     align-items: center;
     cursor: text;
     font-size: 1.5em;
 }
 
-input, label {
+input,
+label {
     display: flex;
     flex-direction: column;
     margin: 0 auto;
 }
 
-#error-container, #success-container {
+#error-container,
+#success-container {
     width: 100%;
     height: 100%;
     justify-items: center;
